+++ livestatus/src/MonitoringCore.h
@@ -71,7 +71,7 @@ public:
                                                    const Contact *contact) = 0;
 
     virtual std::chrono::system_clock::time_point last_logfile_rotation() = 0;
-    virtual uint32_t maxLinesPerLogFile() const = 0;
+    virtual size_t maxLinesPerLogFile() const = 0;
 
     virtual Command find_command(std::string name) const = 0;
     virtual std::vector<Command> commands() const = 0;
