+++ livestatus/src/module.cc
@@ -122,7 +122,7 @@ static vector<ThreadInfo> fl_thread_info;
 static thread_local ThreadInfo *tl_info;
 size_t fl_max_cached_messages = 500000;
 // do never read more than that number of lines from a logfile
-static uint32_t fl_max_lines_per_logfile = 1000000;
+static size_t fl_max_lines_per_logfile = 1000000;
 size_t fl_max_response_size = 100 * 1024 * 1024;  // limit answer to 10 MB
 int g_thread_running = 0;
 static AuthorizationKind fl_service_authorization = AuthorizationKind::loose;
@@ -618,7 +618,7 @@ public:
         return {"", ""};
     }
 
-    uint32_t maxLinesPerLogFile() const override {
+    size_t maxLinesPerLogFile() const override {
         return fl_max_lines_per_logfile;
     }
